+++ readme.md
@@ -31,7 +31,7 @@
 
 ## 1차 리뷰 반영
 
-- [ ] 불필요한 패키지 구조 제거
+- [x] 불필요한 패키지 구조 제거
 - [ ] Car 관련 검증 로직은 Car 안에서만 처리하기
 - [ ] 출력 관련 기능 분리
